+++ src/components/order/CardSection.tsx
@@ -13,7 +13,6 @@ type CardSectionProps = {
     setValue: (value: string) => void;
     error: string;
     onChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
-    onBlur: () => void;
   };
 };
